+++ Engine/src/Engine/GraphicsAPI/VulkanAPI/VulkanRenderPass.cpp
@@ -82,7 +82,7 @@ namespace Vulkan
 
 		VkRenderPassCreateInfo renderPassCreateInfo{};
 		renderPassCreateInfo.sType = VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO;
-		renderPassCreateInfo.attachmentCount = attachments.GetSize();
+		renderPassCreateInfo.attachmentCount = static_cast<uint32_t>(attachments.GetSize());
 		renderPassCreateInfo.pAttachments = attachments.GetFirstPtr();
 		renderPassCreateInfo.subpassCount = 1;
 		renderPassCreateInfo.pSubpasses = &subpassDescription;
